Fix typo in VScaleExpMinusMaxMicrokernelTester
PiperOrigin-RevId: 306568824

+++ test/vscaleexpminusmax-microkernel-tester.h
@@ -38,7 +38,7 @@ class VScaleExpMinusMaxMicrokernelTester {
     return *this;
   }
 
-  inline size_t scale() const {
+  inline float scale() const {
     return this->scale_;
   }
